Move serialize.py
